+++ CHANGELOG.md
@@ -9,12 +9,13 @@
 ### Fixes
 
 - StackTrace parsing on Windows Powershell 5.1 ([#50](https://github.com/getsentry/sentry-powershell/pull/50))
+- Fix out of bounds context for short scripts ([#58](https://github.com/getsentry/sentry-powershell/pull/58))
 
 ### Dependencies
 
-- Bump Dotnet SDK from v4.4.0 to v4.10.2 ([#47](https://github.com/getsentry/sentry-powershell/pull/47), [#51](https://github.com/getsentry/sentry-powershell/pull/51), [#52](https://github.com/getsentry/sentry-powershell/pull/52), [#53](https://github.com/getsentry/sentry-powershell/pull/53))
-  - [changelog](https://github.com/getsentry/sentry-dotnet/blob/main/CHANGELOG.md#4102)
-  - [diff](https://github.com/getsentry/sentry-dotnet/compare/4.4.0...4.10.2)
+- Bump Dotnet SDK from v4.4.0 to v4.12.1 ([#47](https://github.com/getsentry/sentry-powershell/pull/47), [#51](https://github.com/getsentry/sentry-powershell/pull/51), [#52](https://github.com/getsentry/sentry-powershell/pull/52), [#53](https://github.com/getsentry/sentry-powershell/pull/53), [#54](https://github.com/getsentry/sentry-powershell/pull/54), [#57](https://github.com/getsentry/sentry-powershell/pull/57))
+  - [changelog](https://github.com/getsentry/sentry-dotnet/blob/main/CHANGELOG.md#4121)
+  - [diff](https://github.com/getsentry/sentry-dotnet/compare/4.4.0...4.12.1)
 
 ## 0.1.0